changed explore-by-example request
