+++ settings.gradle
@@ -1,2 +1,1 @@
-
 rootProject.name = 'polypheny-ui'
