+++ src/app/views/querying/explore-by-example/explore-by-example.component.ts
@@ -287,7 +287,7 @@ export class ExploreByExampleComponent implements OnInit, OnDestroy {
   }
 
   exploreByExampleEnabled(): boolean {
-      return this._information.getEnabledPlugins().includes('Explore-By-Example');
+      return this._crud.getEnabledPlugins().includes('explore-by-example');
   }
 }
